requirements.txt
celery
fastapi
redis
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-distro
opentelemetry-distro[otlp]
opentelemetry-instrumentation-celery
opentelemetry-instrumentation-fastapi
opentelemetry-propagator-b3
opentelemetry-instrumentation-logging